shanawar/sprint5/requirements.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0
aws-cdk.aws.lambda==1.135.0
aws-cdk.core==1.135.0
aws-cdk.aws_events_targets==1.135.0
aws-cdk.aws_cloudwatch_actions==1.135.0
aws-cdk.aws_dynamodb==1.135.0
aws-cdk.aws_codepipeline==1.135.0
aws-cdk.aws_codepipeline_actions==1.135.0
aws-cdk.pipelines==1.135.0
aws-cdk.aws_apigateway==1.135.0
aws-cdk.aws_amplify==1.135.0
aws-cdk.aws_cognito==1.135.0
aws-cdk.aws_lambda_event_sources==1.135.0
requests
pytest
boto3
